Moving factory_unused_options to new file, removing unwanted lines, correctong bug of candidate scraper type, small nits - #468
