+++ tests/directory_parser/test_directory_parser.py
@@ -22,16 +22,15 @@ def test_init(httpd):
 
     # path_regex to mozdownload -t release -p win32 -v latest
     testpath = urljoin(httpd.get_url(), 'directoryparser/')
-    parser1 = DirectoryParser(testpath)
-    parser1.entries.sort()
+    parser = DirectoryParser(testpath)
+    parser.entries.sort()
     testdir = os.listdir(urljoin(httpd.router.doc_root, 'directoryparser'))
     testdir.sort()
-    assert parser1.entries == testdir
+    assert parser.entries == testdir
 
 
 def test_filter(httpd):
     """Testing the DirectoryParser filter method"""
-
     parser = DirectoryParser(urljoin(httpd.get_url(), 'directoryparser', 'filter/'))
 
     # Get the contents of the folder - dirs and files
@@ -54,7 +53,6 @@ def test_filter(httpd):
 
 
 def test_names_with_spaces(httpd):
-
     parser = DirectoryParser(urljoin(httpd.get_url(), 'directoryparser', 'some spaces/'))
 
     # Get the contents of the folder - dirs and files
