+++ tests/factory/test_factory_scraper.py
@@ -24,7 +24,7 @@ from mozdownload import FactoryScraper
     ('daily', None,
      '2013-10-01-03-02-04-mozilla-central-firefox-27.0a1.en-US.win32.installer.exe', {
          'platform': 'win32',
-    }),
+     }),
 
     # TinderboxScraper
     ('tinderbox', None, '1374583608-mozilla-central-firefox-25.0a1.en-US.win32.installer.exe', {
@@ -38,10 +38,8 @@ from mozdownload import FactoryScraper
          'platform': 'mac64',
      }),
 ])
-
 def test_factory(httpd, tmpdir, scraper_type, builds, fname, arg, mocker):
     """Testing various download scenarios for the factory."""
-
     query_builds_by_revision \
         = mocker.patch('mozdownload.treeherder.Treeherder.query_builds_by_revision')
     base_kwargs = {
